requirements.txt
# Core retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
sentence-transformers>=2.7.0
transformers>=4.40.0
numpy>=1.26.0

# Data handling
pandas>=2.2.0
datasets>=2.19.0

# Evaluation & reporting
scikit-learn>=1.4.0
tabulate>=0.9.0
tqdm>=4.66.0
